Implement reward distribution (#25)
* replace NBSP with space

* fix core logic

- store only sdk.Coins when storing a reward in kvstore
- fix store methods
- delete a staking when all coins has been unstaked and no rewards left
- emit events in keeper method rather than in msg handler

* revert back parameter order of store methods

* revert back old Reward struct

* rename and simplify Denoms method

* drop legacy amino dependency and replace simd with farmingd

* implement Stakings query endpoint and minor fixes

- rename key getter methods
- unify variable/parameter namings

* fix Unstake logic

* make tx commands work

* change time format

* move plan-related logics into plan.go

* fix Rewards endpoint logic and minor bugs

* use sdk.BigEndianToUint64 instead of encoding/binary

* fix query endpoint bugs

* improve stake/unstake cli commands

* use test suite for testing

* add tests and update sentinel errors

* make logics executed in EndBlock not BeginBlock

* implement epoch-related functions

* implement ProcessQueuedCoins and add test helper

* add and refactor tests

* implement reward distribution logic

also make Create*Plan to return the plan

* apply PR review suggestions

- remove code for debugging purpose
- accumulate reward coins, not overwrite
- only update staking object when there is a change

+++ x/farming/keeper/staking.go
@@ -3,8 +3,10 @@ package keeper
 import (
 	"encoding/binary"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	gogotypes "github.com/gogo/protobuf/types"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/tendermint/farming/x/farming/types"
 )
@@ -111,7 +113,9 @@ func (k Keeper) DeleteStaking(ctx sdk.Context, staking types.Staking) {
 	store := ctx.KVStore(k.storeKey)
 	store.Delete(types.GetStakingKey(staking.Id))
 	store.Delete(types.GetStakingByFarmerIndexKey(staking.GetFarmer()))
-	k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, staking.StakingCoinDenoms())
+	if denoms := staking.StakingCoinDenoms(); len(denoms) > 0 {
+		k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, denoms)
+	}
 }
 
 // DeleteStakingCoinDenomsIndex removes an staking for the staking mapper store.
@@ -175,22 +179,21 @@ func (k Keeper) ReleaseStakingCoins(ctx sdk.Context, farmer sdk.AccAddress, unst
 }
 
 // Stake stores staking coins to queued coins and it will be processed in the next epoch.
-func (k Keeper) Stake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) error {
+func (k Keeper) Stake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) (types.Staking, error) {
 	if err := k.ReserveStakingCoins(ctx, farmer, amount); err != nil {
-		return err
+		return types.Staking{}, err
 	}
 
 	params := k.GetParams(ctx)
 
 	staking, found := k.GetStakingByFarmer(ctx, farmer)
 	if !found {
-		farmingFeeCollectorAcc, err := sdk.AccAddressFromBech32(params.FarmingFeeCollector)
-		if err != nil {
-			return err
-		}
+		// We don't need to check error here because it is already checked
+		// in Params validation.
+		farmingFeeCollectorAcc, _ := sdk.AccAddressFromBech32(params.FarmingFeeCollector)
 
 		if err := k.bankKeeper.SendCoins(ctx, farmer, farmingFeeCollectorAcc, params.StakingCreationFee); err != nil {
-			return err
+			return types.Staking{}, sdkerrors.Wrap(types.ErrFeeCollectionFailure, err.Error())
 		}
 
 		staking = k.NewStaking(ctx, farmer)
@@ -208,18 +211,18 @@ func (k Keeper) Stake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) 
 		),
 	})
 
-	return nil
+	return staking, nil
 }
 
 // Unstake unstakes an amount of staking coins from the staking reserve account.
-func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) error {
+func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) (types.Staking, error) {
 	staking, found := k.GetStakingByFarmer(ctx, farmer)
 	if !found {
-		return types.ErrStakingNotExists
+		return types.Staking{}, types.ErrStakingNotExists
 	}
 
 	if err := k.ReleaseStakingCoins(ctx, farmer, amount); err != nil {
-		return err
+		return types.Staking{}, err
 	}
 
 	prevDenoms := staking.StakingCoinDenoms()
@@ -239,6 +242,7 @@ func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins
 
 	// Remove the Staking object from the kvstore when all coins has been unstaked
 	// and there's no rewards left.
+	// TODO: find more efficient way to check if the farmer has no rewards
 	if staking.StakedCoins.IsZero() && staking.QueuedCoins.IsZero() && len(k.GetRewardsByFarmer(ctx, farmer)) == 0 {
 		k.DeleteStaking(ctx, staking)
 	} else {
@@ -256,7 +260,9 @@ func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins
 			}
 		}
 
-		k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, removedDenoms)
+		if len(removedDenoms) > 0 {
+			k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, removedDenoms)
+		}
 	}
 
 	ctx.EventManager().EmitEvents(sdk.Events{
@@ -267,5 +273,18 @@ func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins
 		),
 	})
 
-	return nil
+	// We're returning a Staking even if it has deleted.
+	return staking, nil
+}
+
+// ProcessQueuedCoins moves queued coins into staked coins.
+func (k Keeper) ProcessQueuedCoins(ctx sdk.Context) {
+	k.IterateAllStakings(ctx, func(staking types.Staking) (stop bool) {
+		if !staking.QueuedCoins.IsZero() {
+			staking.StakedCoins = staking.StakedCoins.Add(staking.QueuedCoins...)
+			staking.QueuedCoins = sdk.NewCoins()
+			k.SetStaking(ctx, staking)
+		}
+		return false
+	})
 }
